% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avert-package.R, R/pkg_avert.R
\docType{package}
\name{avert-package}
\alias{avert-package}
\title{avert: Avertable Burden Decision Aid}
\description{
Package containing the Shiny application and supporting
workflows for generating avertable burden reports.

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An in development library for predicting the burden due to mental disorders that is theoretically avertable with existing healthcare interventions. This library is being developed with the ready4 framework (https://ready4-dev.github.io/ready4/). This software has been made publicly available as part of the process of testing and documenting the library. For further information email matthew.hamilton2@monash.edu.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://avertableburden.github.io/avert/}
  \item \url{https://github.com/avertableburden/avert}
  \item \url{https://ready4-dev.github.io/ready4/}
}


Useful links:
\itemize{
  \item \url{https://avertableburden.github.io/avert/}
  \item \url{https://github.com/avertableburden/avert}
  \item \url{https://ready4-dev.github.io/ready4/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton2@monash.edu} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

}
\keyword{internal}
