% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_app_server.R
\name{avert_server}
\alias{avert_server}
\title{Server Logic for the Avertable Burden App}
\usage{
avert_server(input, output, session, areas_chr = make_avert_areas())
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{areas_chr}{Character vector of available geographic areas.}
}
\value{
No return value. Called for side effects.
}
\description{
Implements server-side processing for report generation,
previewing and export.
}
\keyword{internal}
