% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_app_ui.R
\name{avert_ui}
\alias{avert_ui}
\title{User Interface for the Avertable Burden App}
\usage{
avert_ui(areas_chr = make_avert_areas())
}
\arguments{
\item{areas_chr}{Character vector of available geographic areas.}
}
\value{
A Shiny UI object.
}
\description{
Creates the user interface for the Avertable Burden
report-generation application.
}
\keyword{internal}
