% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_input_data}
\alias{get_input_data}
\title{Get input data}
\usage{
get_input_data()
}
\value{
Parameters (an output object of multiple potential types)
}
\description{
get_input_data() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get input data. The function returns Parameters (an output object of multiple potential types).
}
