% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_parameter_value}
\alias{get_parameter_value}
\title{Get parameter value}
\usage{
get_parameter_value(parameter_1L_chr = "first", parameters_ls = NULL)
}
\arguments{
\item{parameter_1L_chr}{Parameter (a character vector of length one), Default: 'first'}

\item{parameters_ls}{Parameters (a list), Default: NULL}
}
\value{
Parameter (an output object of multiple potential types)
}
\description{
get_parameter_value() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get parameter value. The function returns Parameter (an output object of multiple potential types).
}
