% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_project_repos}
\alias{get_project_repos}
\title{Get project repositories}
\usage{
get_project_repos()
}
\value{
Repositories (a character vector)
}
\description{
get_project_repos() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get project repositories. The function returns Repositories (a character vector).
}
