% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_run_app.R
\name{run_avert_app}
\alias{run_avert_app}
\title{Launch the Avertable Burden App}
\usage{
run_avert_app(launch_browser = TRUE)
}
\arguments{
\item{launch_browser}{Logical indicating whether the app should
automatically open in a browser.}
}
\value{
Invisibly returns the launched application.
}
\description{
Starts the Avertable Burden Shiny application.
}
\examples{
\dontrun{
run_avert_app()
}

}
